% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM2YOLO.R
\name{EM2YOLO}
\alias{EM2YOLO}
\title{Convert an EventMeasure Database to YOLO annotation format}
\usage{
EM2YOLO(
  EMDB,
  pic.dir,
  group.name,
  species.index,
  img.height,
  img.width,
  YOLO.dir,
  remove.anno = NULL
)
}
\arguments{
\item{EMDB}{The file path to the EventMeasure database. This may be a modified version of the EventMeasure export created with RenameSpecies()}

\item{pic.dir}{The path to the directory containg the images reffered to in the specied EventMeasure database}

\item{group.name}{The identifier to give this set of images. This must be a character string and will form the filename for the images copied in to the YOLO image directories. Image filenames will be 'NUMBER_group.name.jpg'.}

\item{species.index}{A dataframe containing each unique value in the EventMeasure database 'Species' column and an associated index value. This should be created with the EMCreateYAML() function so as to be consistent with the YAML configuration file for the YOLO model.}

\item{img.height}{The height (in pixels) of annotated images (required to transform bounding box coordinates to YOLO format).}

\item{img.width}{The width (in pixels) of annotated images (required to transform bounding box coordinates to YOLO format).}

\item{YOLO.dir}{The path to the directory containing the YOLO training data. This directory needs to have a specific structure, containing two subdirectories named 'images' and 'labels'. Each of these must then have three subdirectories named 'Train', 'Val' and 'Test'.}

\item{remove.anno}{An optional character vector containing the name of any species that should not be included in the YOLO annotations. The species names must appear exactly as they do in the 'Species' Column of the EventMeasure databse. This can be useful when, for example, multiple species are annotated but a detector is required for only a small number of those.}
}
\description{
This function takes an EventMeasure database export and converts it to the
YOLO annotation format. Images are randomly sampled and copied in to the 
relevant YOLO training/validation/testing directories with an 80/10/10 split.
.
}
\examples{
EM2YOLO(EMDB = 'path/to/EMdatabase.txt',
        pic.dir = 'path/to/EM/images',
        group.name = 'Set1',
        species.index = sptable,
        img.height = 1080,
        img.width = 1920,
        YOLO.dir = path/to/YOLO/data,
        remove.anno = c('sp1', 'sp2'))
}
